# docs/詳細設計/api/openapi.yaml
openapi: 3.1.0
info:
  title: Character Taste Lab API
  version: 1.0.0
  description: >-
    DD-03の機械可読契約。個人データの操作はcookie sessionと
    X-CSRF-Tokenを使い、全mutationはIdempotency-Keyを必須とする。
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      security: []
      parameters:
        - name: search
          in: query
          schema: { type: string, maxLength: 32 }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ListLimit100'
      responses:
        '200':
          description: Active public users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicUserListResponse' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Users]
      operationId: createUser
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Pending user and one-time access key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateUserResponse' }
        default: { $ref: '#/components/responses/Error' }
  /users/{userId}/activate:
    post:
      tags: [Users]
      operationId: activateUser
      security: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [accessKey]
              properties:
                accessKey: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200': { $ref: '#/components/responses/ObjectData' }
        default: { $ref: '#/components/responses/Error' }
  /sessions:
    post:
      tags: [Sessions]
      operationId: createSession
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '200':
          description: Session created; __Host-session is set
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Sessions]
      operationId: deleteSession
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OptionalCsrfToken'
      responses:
        '204': { description: Session revoked and cookie cleared }
        default: { $ref: '#/components/responses/Error' }
  /me:
    get:
      tags: [Sessions]
      operationId: getMe
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        default: { $ref: '#/components/responses/Error' }
  /account/key-rotation:
    post:
      tags: [Account]
      operationId: rotateAccessKey
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentAccessKey]
              properties:
                currentAccessKey: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: One-time new key; all sessions revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RotateKeyResponse' }
        default: { $ref: '#/components/responses/Error' }
...
  /account:
    delete:
      tags: [Account]
      operationId: deleteAccount
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [usernameConfirmation]
              properties:
                usernameConfirmation: { type: string, minLength: 1, maxLength: 32 }
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        default: { $ref: '#/components/responses/Error' }
  /works:
    get:
      tags: [Catalog]
      operationId: listWorks
      parameters:
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ListLimit100'
      responses:
        '200': { $ref: '#/components/responses/ListData' }
        default: { $ref: '#/components/responses/Error' }
...
  /entries:
    get:
      tags: [Entries]
      operationId: listEntries
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: registrationType
          in: query
          schema: { $ref: '#/components/schemas/RegistrationType' }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ListLimit100'
      responses:
        '200': { $ref: '#/components/responses/ListData' }
        default: { $ref: '#/components/responses/Error' }
...
  /entries/{entryId}:
    get:
      tags: [Entries]
      operationId: getEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
      responses:
        '200': { $ref: '#/components/responses/ObjectData' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Entries]
      operationId: deleteEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '204': { description: Draft deleted }
        default: { $ref: '#/components/responses/Error' }
...
  /entries/{entryId}/reanalysis:
    post:
      tags: [Entries]
      operationId: reanalyzeEntry
      parameters:
        - $ref: '#/components/parameters/EntryId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntryReanalysisRequest' }
      responses:
        '202': { $ref: '#/components/responses/ObjectData' }
        default: { $ref: '#/components/responses/Error' }
...
  /profile:
    get:
      tags: [Profile]
      operationId: getProfile
      responses:
        '200': { $ref: '#/components/responses/ObjectData' }
        default: { $ref: '#/components/responses/Error' }
...
  /profile/graph:
    get:
      tags: [Graph]
      operationId: getProfileGraph
      parameters:
        - name: profileSnapshotId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: projectionVersion
          in: query
          schema: { type: string }
        - name: detail
          in: query
          schema: { type: string, enum: [summary, standard, expanded], default: standard }
        - $ref: '#/components/parameters/Cursor'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 500 }
      responses:
        '200':
          description: Graph projection page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GraphProjectionPageResponse' }
        default: { $ref: '#/components/responses/Error' }
...
  /generation-requests:
    post:
      tags: [Generation]
      operationId: createGenerationRequest
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGenerationRequest' }
      responses:
        '201': { $ref: '#/components/responses/ObjectData' }
        default: { $ref: '#/components/responses/Error' }
...
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      operationId: getJob
      security:
        - sessionCookie: []
        - deletionStatusToken: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobResponseEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /jobs/{jobId}/retry:
    post:
      tags: [Jobs]
      operationId: retryJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
    deletionStatusToken:
      type: apiKey
      in: header
      name: X-Deletion-Status-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 2048 }
    ListLimit100:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    UserId: { name: userId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ObjectData:
      description: Object data envelope; the operation description defines the concrete DTO
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ObjectDataResponse' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
    IsoDateTime: { type: string, format: date-time }
...
    PublicUserListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/PublicUser' } }
        meta: { $ref: '#/components/schemas/PageResponseMeta' }
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, consentVersion, turnstileToken]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        consentVersion: { type: string, minLength: 1, maxLength: 50 }
        turnstileToken: { type: string, minLength: 1, maxLength: 4096 }
    CreateUserResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [user, accessKey, expiresAt]
          properties:
            user:
              type: object
              additionalProperties: false
              required: [id, username, status]
              properties:
                id: { $ref: '#/components/schemas/Uuid' }
                username: { type: string }
                status: { const: pending }
            accessKey: { $ref: '#/components/schemas/Uuid' }
            expiresAt: { $ref: '#/components/schemas/IsoDateTime' }
        meta: { $ref: '#/components/schemas/RequestMeta' }
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [userId, accessKey, turnstileToken]
      properties:
        userId: { $ref: '#/components/schemas/Uuid' }
        accessKey: { $ref: '#/components/schemas/Uuid' }
        turnstileToken: { type: string, minLength: 1, maxLength: 4096 }
    SessionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [user, csrfToken, expiresAt]
          properties:
            user: { $ref: '#/components/schemas/PublicUser' }
            csrfToken: { type: string, pattern: '^[0-9a-f]{64}$' }
            expiresAt: { $ref: '#/components/schemas/IsoDateTime' }
            capacity: { $ref: '#/components/schemas/CapacityState' }
        meta: { $ref: '#/components/schemas/RequestMeta' }
...
    RegistrationType: { type: string, enum: [existing, customized_existing, original] }
...